requirements.txt
certifi==2025.11.12
charset-normalizer==3.4.4
idna==3.11
numpy==2.3.5
oauthlib==3.3.1
pillow==12.0.0
python-dotenv==1.2.1
requests==2.32.5
requests-oauthlib==2.0.0
tweepy==4.16.0
urllib3==2.5.0
